+++ src/bittytax/conv/parsers/binance.py
@@ -81,6 +81,7 @@ BASE_ASSETS = [
 
 TRADINGPAIR_TO_QUOTE_ASSET = {
     "ADAEUR": "EUR",
+    "ENAEUR": "EUR",
     "GALAEUR": "EUR",
     "LUNAEUR": "EUR",
     "THETAEUR": "EUR",
@@ -256,10 +257,17 @@ def parse_binance_deposits_withdrawals_crypto(
 
 
 def parse_binance_deposits_withdrawals_cash(
-    data_row: "DataRow", _parser: DataParser, **kwargs: Unpack[ParserArgs]
+    data_row: "DataRow", parser: DataParser, **kwargs: Unpack[ParserArgs]
 ) -> None:
     row_dict = data_row.row_dict
-    data_row.timestamp = DataParser.parse_timestamp(row_dict["Date(UTC)"])
+
+    timestamp_hdr = parser.args[0].group(1)
+    utc_offset = parser.args[0].group(2)
+
+    if utc_offset == "UTCnull":
+        utc_offset = "UTC"
+
+    data_row.timestamp = DataParser.parse_timestamp(f"{row_dict[timestamp_hdr]} {utc_offset}")
 
     if row_dict["Status"] != "Successful":
         return
@@ -792,7 +800,7 @@ DataParser(
     ParserType.EXCHANGE,
     "Binance Deposits/Withdrawals",
     [
-        "Date(UTC)",
+        lambda c: re.match(r"(^Date\((UTC|UTCnull|UTC[-+]\d{1,2})\))", c),
         "Coin",
         "Amount",
         "Status",