+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Coinbase parser: fixed missing "Fee Quantity" when zero.
 - KuCoin parser: use UTC offset from Time header for timestamp.
 - Accounting/Conversion tool: catch exception when importing a Google Sheet.
+- Binance parser: use UTC offset from Date(UTC) header if present. ([#302](https://github.com/BittyTax/BittyTax/issues/302))
 ### Added
 - Accounting tool: new PDF report format.
 - Config: added classic_report parameter to switch to legacy PDF report format.